Code cleanup
* Remove ctor on SequenceReader to make it compatible with netcoreapp3.0 binary
* Avoid using Utilities extension in BufferWriter
* Move NilFormatter to MessagePack.Formatters
