+++ src/MessagePack/Nil.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Buffers;
 
 namespace MessagePack
 {
@@ -25,52 +24,6 @@ namespace MessagePack
         public override string ToString()
         {
             return "()";
-        }
-    }
-}
-
-namespace MessagePack.Formatters
-{
-    public class NilFormatter : IMessagePackFormatter<Nil>
-    {
-        public static readonly IMessagePackFormatter<Nil> Instance = new NilFormatter();
-
-        NilFormatter()
-        {
-
-        }
-
-        public void Serialize(ref MessagePackWriter writer, Nil value, IFormatterResolver typeResolver)
-           
-        {
-            writer.WriteNil();
-        }
-
-        public Nil Deserialize(ref MessagePackReader reader, IFormatterResolver typeResolver)
-        {
-            return reader.ReadNil();
-        }
-    }
-
-    // NullableNil is same as Nil.
-    public class NullableNilFormatter : IMessagePackFormatter<Nil?>
-    {
-        public static readonly IMessagePackFormatter<Nil?> Instance = new NullableNilFormatter();
-
-        NullableNilFormatter()
-        {
-
-        }
-
-        public void Serialize(ref MessagePackWriter writer, Nil? value, IFormatterResolver typeResolver)
-           
-        {
-            writer.WriteNil();
-        }
-
-        public Nil? Deserialize(ref MessagePackReader reader, IFormatterResolver typeResolver)
-        {
-            return reader.ReadNil();
         }
     }
 }
