+++ src/MessagePack/SequenceReaderExtensions.cs
@@ -55,24 +55,6 @@ namespace System.Buffers
         }
 
         /// <summary>
-        /// Reads an <see cref="sbyte"/> from the next position in the sequence.
-        /// </summary>
-        /// <param name="reader">The reader to read from.</param>
-        /// <param name="value">Receives the value read.</param>
-        /// <returns><c>true</c> if there was another byte in the sequence; <c>false</c> otherwise.</returns>
-        public static bool TryRead(ref this SequenceReader<byte> reader, out sbyte value)
-        {
-            if (TryRead(ref reader, out byte byteValue))
-            {
-                value = unchecked((sbyte)byteValue);
-                return true;
-            }
-
-            value = default;
-            return false;
-        }
-
-        /// <summary>
         /// Reads an <see cref="Int16"/> as little endian.
         /// </summary>
         /// <returns>False if there wasn't enough data for an <see cref="Int16"/>.</returns>
@@ -98,22 +80,6 @@ namespace System.Buffers
             }
 
             return TryReadReverseEndianness(ref reader, out value);
-        }
-
-        /// <summary>
-        /// Reads an <see cref="UInt16"/> as big endian.
-        /// </summary>
-        /// <returns>False if there wasn't enough data for an <see cref="UInt16"/>.</returns>
-        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out ushort value)
-        {
-            if (TryReadBigEndian(ref reader, out short shortValue))
-            {
-                value = unchecked((ushort)shortValue);
-                return true;
-            }
-
-            value = default;
-            return false;
         }
 
         private static bool TryReadReverseEndianness(ref SequenceReader<byte> reader, out short value)
@@ -155,22 +121,6 @@ namespace System.Buffers
             return TryReadReverseEndianness(ref reader, out value);
         }
 
-        /// <summary>
-        /// Reads an <see cref="UInt32"/> as big endian.
-        /// </summary>
-        /// <returns>False if there wasn't enough data for an <see cref="UInt32"/>.</returns>
-        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out uint value)
-        {
-            if (TryReadBigEndian(ref reader, out int intValue))
-            {
-                value = unchecked((uint)intValue);
-                return true;
-            }
-
-            value = default;
-            return false;
-        }
-
         private static bool TryReadReverseEndianness(ref SequenceReader<byte> reader, out int value)
         {
             if (reader.TryRead(out value))
@@ -208,22 +158,6 @@ namespace System.Buffers
             }
 
             return TryReadReverseEndianness(ref reader, out value);
-        }
-
-        /// <summary>
-        /// Reads an <see cref="UInt64"/> as big endian.
-        /// </summary>
-        /// <returns>False if there wasn't enough data for an <see cref="UInt64"/>.</returns>
-        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out ulong value)
-        {
-            if (TryReadBigEndian(ref reader, out long longValue))
-            {
-                value = unchecked((ulong)longValue);
-                return true;
-            }
-
-            value = default;
-            return false;
         }
 
         private static bool TryReadReverseEndianness(ref SequenceReader<byte> reader, out long value)
