+++ src/MessagePack/SequenceReaderExtensions2.cs
@@ -1,0 +1,75 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Buffers
+{
+    internal static partial class SequenceReaderExtensions
+    {
+        /// <summary>
+        /// Reads an <see cref="sbyte"/> from the next position in the sequence.
+        /// </summary>
+        /// <param name="reader">The reader to read from.</param>
+        /// <param name="value">Receives the value read.</param>
+        /// <returns><c>true</c> if there was another byte in the sequence; <c>false</c> otherwise.</returns>
+        public static bool TryRead(ref this SequenceReader<byte> reader, out sbyte value)
+        {
+            if (reader.TryRead(out byte byteValue))
+            {
+                value = unchecked((sbyte)byteValue);
+                return true;
+            }
+
+            value = default;
+            return false;
+        }
+
+        /// <summary>
+        /// Reads an <see cref="UInt16"/> as big endian.
+        /// </summary>
+        /// <returns>False if there wasn't enough data for an <see cref="UInt16"/>.</returns>
+        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out ushort value)
+        {
+            if (reader.TryReadBigEndian(out short shortValue))
+            {
+                value = unchecked((ushort)shortValue);
+                return true;
+            }
+
+            value = default;
+            return false;
+        }
+
+        /// <summary>
+        /// Reads an <see cref="UInt32"/> as big endian.
+        /// </summary>
+        /// <returns>False if there wasn't enough data for an <see cref="UInt32"/>.</returns>
+        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out uint value)
+        {
+            if (reader.TryReadBigEndian(out int intValue))
+            {
+                value = unchecked((uint)intValue);
+                return true;
+            }
+
+            value = default;
+            return false;
+        }
+
+        /// <summary>
+        /// Reads an <see cref="UInt64"/> as big endian.
+        /// </summary>
+        /// <returns>False if there wasn't enough data for an <see cref="UInt64"/>.</returns>
+        public static bool TryReadBigEndian(ref this SequenceReader<byte> reader, out ulong value)
+        {
+            if (reader.TryReadBigEndian(out long longValue))
+            {
+                value = unchecked((ulong)longValue);
+                return true;
+            }
+
+            value = default;
+            return false;
+        }
+    }
+}
